refactor: 장소 간 거리, 시간 계산기가 자동 등록되도록 변경
