+++ backend/spring-routie/src/main/java/routie/place/infrastructure/distancecalculator/DistanceCalculatorComposite.java
@@ -1,25 +1,17 @@
-package routie.place.infrastructure;
+package routie.place.infrastructure.distancecalculator;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Primary;
-import org.springframework.stereotype.Component;
 import routie.place.domain.DistanceCalculator;
 import routie.place.domain.MovingStrategy;
 import routie.place.domain.Place;
 
-@Primary
-@Component
 public class DistanceCalculatorComposite implements DistanceCalculator {
 
     private final List<DistanceCalculator> distanceCalculators;
 
-    public DistanceCalculatorComposite(
-            @Qualifier("fakeDrivingDistanceCalculator") final DistanceCalculator fakeDrivingDistanceCalculator
-    ) {
-        this.distanceCalculators = new ArrayList<>();
-        this.distanceCalculators.add(fakeDrivingDistanceCalculator);
+    public DistanceCalculatorComposite(final List<DistanceCalculator> drivingDistanceCalculators) {
+        this.distanceCalculators = new ArrayList<>(drivingDistanceCalculators);
     }
 
     @Override
@@ -38,5 +30,4 @@ public class DistanceCalculatorComposite implements DistanceCalculator {
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("지원하지 않는 이동 방식입니다: " + strategy));
     }
-
 }
